server/pyproject.toml
[tool.poetry]
name = "server"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^5.0.4"
djangorestframework = "^3.15.1"
djangorestframework-simplejwt = "^5.3.1"
psycopg2 = "^2.9.9"
django-cors-headers = "^4.3.1"
pillow = "^10.3.0"
qrcode = "^7.4.2"
django-haystack = "^3.2.1"
pyproj = "^3.6.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
